// packages/@vuepress/core/src/types/app.ts
import type { Markdown } from '@vuepress/markdown'
import type { ThemeApi } from './theme-api'
import type { BundlerFunc } from './bundler'
import type { Page } from './page'
import type { Plugin, PluginOptions, PluginConfig } from './plugin'
import type { PluginApi } from './plugin-api'
import type { VuepressConfig } from './config'

/**
 * Vuepress app
 */
export interface App {
  // created
  version: string
  options: AppOptions
  dir: AppDir
  env: AppEnv
  markdown: Markdown
  pluginApi: PluginApi
  use: <T extends PluginOptions>(
    plugin: Plugin<T> | string,
    config?: Partial<T>
  ) => void
  useByConfig: <T extends PluginOptions>(pluginConfig: PluginConfig<T>) => void
  init: () => Promise<void>
  prepare: () => Promise<void>
  dev: <T = void>() => ReturnType<BundlerFunc<T>>
  build: <T = void>() => ReturnType<BundlerFunc<T>>

  // initialized
  writeTemp: AppWriteTemp
  themeApi: ThemeApi

  // prepared
  pages: Page[]
}

/**
 * Vuepress app options
 */
export type AppOptions = Required<VuepressConfig>

/**
 * Directory util function
 */
export type AppDirFunction = (...args: string[]) => string

/**
 * Directory utils
 */
export interface AppDir {
  temp: AppDirFunction
  source: AppDirFunction
  dest: AppDirFunction
  client: AppDirFunction
}

/**
 * Environment flags
 */
export interface AppEnv {
  isProd: boolean
  isTest: boolean
  isDev: boolean
  isDebug: boolean
  nodeEnv: string | undefined
}

/**
 * Write temp file util
 */
export type AppWriteTemp = (file: string, content: string) => Promise<string>